# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /field/assignments:
    get:
      operationId: listFieldAssignments
      tags:
        - field
      summary: List field assignments
      description: Returns all relayed field assignments with their evidence
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum:
              - assigned
              - in_progress
              - completed
              - cancelled
      responses:
        '200':
          description: Assignments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FieldAssignmentSync'
    put:
      operationId: pushFieldAssignments
      tags:
        - field
      summary: Push field assignments from desktop
      description: >-
        Bulk-upserts assignments into the relay. Snapshot fields are always refreshed; mutable
        service fields use last-write-wins by updatedAt. Evidence is append-only by id.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PushFieldAssignmentsRequest'
      responses:
        '200':
          description: Push result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PushFieldAssignmentsResult'
  '/field/assignments/{id}':
    patch:
      operationId: updateFieldAssignment
      tags:
        - field
      summary: Update a field assignment (mobile)
      description: Records service progress from the field agent
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateFieldAssignmentRequest'
      responses:
        '200':
          description: Updated assignment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FieldAssignmentSync'
        '404':
          description: Assignment not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
  '/field/assignments/{id}/evidence':
    post:
      operationId: addFieldEvidence
      tags:
        - field
      summary: Attach service evidence (mobile)
      description: >-
        Adds a photo + GPS evidence record to an assignment. Idempotent by client-generated evidence
        id so offline queues can safely retry.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddFieldEvidenceRequest'
      responses:
        '200':
          description: Updated assignment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FieldAssignmentSync'
        '404':
          description: Assignment not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
...
components:
...
  schemas:
...
    ErrorMessage:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    FieldEvidenceSync:
      type: object
      properties:
        id:
          type: string
        photoDataUrl:
          type: string
        latitude:
          type:
            - number
            - 'null'
        longitude:
          type:
            - number
            - 'null'
        accuracyMeters:
          type:
            - number
            - 'null'
        capturedAt:
          type: string
        note:
          type: string
      required:
        - id
        - photoDataUrl
        - latitude
        - longitude
        - accuracyMeters
        - capturedAt
        - note
    FieldAssignmentSync:
      type: object
      properties:
        id:
          type: string
        noticeId:
          type: string
        assigneeName:
          type: string
        instructions:
          type: string
        status:
          type: string
          enum:
            - assigned
            - in_progress
            - completed
            - cancelled
        serviceMethod:
          type:
            - string
            - 'null'
          enum:
            - personal
            - substitute
            - post_and_mail
            - other
            - null
        completedAt:
          type:
            - string
            - 'null'
        serverNotes:
          type: string
        tenantNames:
          type: array
          items:
            type: string
        propertyAddress:
          type: string
        unit:
          type: string
        noticeType:
          type: string
        deadlineDate:
          type:
            - string
            - 'null'
        totalAmountCents:
          type:
            - integer
            - 'null'
        source:
          description: >-
            Origin of the underlying tenant/ledger data (e.g. "buildium").
            Omitted or null for manually entered data.
          type:
            - string
            - 'null'
        evidence:
          type: array
          items:
            $ref: '#/components/schemas/FieldEvidenceSync'
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - noticeId
        - assigneeName
        - instructions
        - status
        - serviceMethod
        - completedAt
        - serverNotes
        - tenantNames
        - propertyAddress
        - unit
        - noticeType
        - deadlineDate
        - totalAmountCents
        - evidence
        - createdAt
        - updatedAt
    PushFieldAssignmentsRequest:
      type: object
      properties:
        assignments:
          type: array
          items:
            $ref: '#/components/schemas/FieldAssignmentSync'
      required:
        - assignments
    PushFieldAssignmentsResult:
      type: object
      properties:
        pushed:
          type: integer
      required:
        - pushed
    UpdateFieldAssignmentRequest:
      type: object
      properties:
        status:
          type: string
          enum:
            - assigned
            - in_progress
            - completed
            - cancelled
        serviceMethod:
          type:
            - string
            - 'null'
          enum:
            - personal
            - substitute
            - post_and_mail
            - other
            - null
        completedAt:
          type:
            - string
            - 'null'
        serverNotes:
          type: string
        updatedAt:
          type: string
      required:
        - updatedAt
    AddFieldEvidenceRequest:
      type: object
      properties:
        id:
          type: string
        photoDataUrl:
          type: string
        latitude:
          type:
            - number
            - 'null'
        longitude:
          type:
            - number
            - 'null'
        accuracyMeters:
          type:
            - number
            - 'null'
        capturedAt:
          type: string
        note:
          type: string
      required:
        - id
        - photoDataUrl
        - capturedAt
...